renamed:    src/front/js/pages/info.js -> src/front/js/component/info.js 	modified:   src/front/js/layout.js
