+++ src/front/js/component/info.js
